New approach to trigger the compilation of our passes.
Instead of using alwayslink = True in the corresponding BUILD files in
XLA and LLVM, remove those and use alwayslink = True in the library
that we generate where we compile our passes.

This approach "works", because at least tfserving builds and also
startup does not fail. However it segfaults when running the client.
However, this issue is probably unrelated to this commit...